Move to post-compile weave to fix conflict with annotation processors
